requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0
duckdb>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
